refactor: make minor changes
