+++ pom.xml
@@ -67,7 +67,6 @@
 		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
-			<optional>true</optional>
 		</dependency>
 		<!-- Testing -->
 		<dependency>
