--- src/main/java/org/example/learningsystem/dto/ErrorResponseDto.java
@@ -1,9 +1,0 @@
-package org.example.learningsystem.dto;
-
-import java.time.LocalDateTime;
-
-public record ErrorResponseDto(String message, Integer code, LocalDateTime timestamp) {
-    public ErrorResponseDto(String message, Integer code) {
-        this(message, code, LocalDateTime.now());
-    }
-}
