+++ src/main/java/org/example/learningsystem/exception/response/ErrorResponse.java
@@ -1,0 +1,9 @@
+package org.example.learningsystem.exception.response;
+
+import java.time.LocalDateTime;
+
+public record ErrorResponse(String message, LocalDateTime timestamp) {
+    public ErrorResponse(String message) {
+        this(message, LocalDateTime.now());
+    }
+}
